move defineDecorator into its own module to resolve circular dependency issues
