+++ define-decorator.d.ts
@@ -1,0 +1,1 @@
+export declare function defineDecorator(proto: Object|Function, key: string, decorator:Function): void;
